HARMONY-1442: Clean up debug statements and commented out code.
