+++ tasks/query-cmr/test/work.ts
@@ -4,33 +4,12 @@ import { describe, it } from 'mocha';
 import { hookDoWork } from './helpers/work';
 import CmrStacCatalog from '../app/stac/cmr-catalog';
 import { resolve } from '../../../app/util/url';
-import { FileStore } from '../../../app/util/object-store/file-store';
-import * as objStore from '../../../app/util/object-store';
-import * as sinon from 'sinon';
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-// import * as objStore from '../../../test/helpers/object-store';
+import { hookMockS3 } from '../../../test/helpers/object-store';
+import { defaultObjectStore } from '@harmony/util/object-store';
 
 describe('doWork', function () {
+  hookMockS3();
   describe('main', function () {
-
-    before(async function () {
-      // Stub cmr fetch post to return the contents of queries
-
-      // Stub object-store calls to store/get query params
-      this.store = new FileStore();
-      // this.uploadStub = sinon.stub(this.store, 'upload');
-      // this.downloadStub = sinon.stub(this.store, 'getObject').returns(Promise.resolve('{"collection_concept_id": "C001-TEST"}'));
-      this.defaultStoreStub = sinon.stub(objStore, 'defaultObjectStore').returns(this.store);
-      this.protocolStub = sinon.stub(objStore, 'objectStoreForProtocol').returns(this.store);
-    });
-    after(function () {
-      this.defaultStoreStub.restore();
-      // this.uploadStub.restore();
-      // this.downloadStub.restore();
-      this.protocolStub.restore();
-      delete this.store;
-    });
     describe('when the output directory exists', function () {
       const totalItemsSize = 1.0;
       const workRequest: QueryCmrRequest = {
@@ -43,8 +22,7 @@ describe('doWork', function () {
       );
 
       it('outputs the result data to catalog.json in the directory', async function () {
-        const catalog = await this.store.getObject(resolve(workRequest.outputDir, 'catalog0.json'));
-        // console.log(`Catalog is ${catalog}`);
+        const catalog = await defaultObjectStore().getObject(resolve(workRequest.outputDir, 'catalog0.json'));
         expect(JSON.parse(catalog).description).to.equal('done');
       });
     });
