+++ test/helpers/object-store.ts
@@ -1,143 +1,7 @@
-// import { stub, SinonStub } from 'sinon';
-// import fs from 'fs';
-// import mockAws from 'mock-aws-s3';
 import * as tmp from 'tmp';
-// import { S3ObjectStore, BucketParams } from '../../app/util/object-store/s3-object-store';
-// import { PutObjectCommandOutput, S3Client } from '@aws-sdk/client-s3';
-
 import { FileStore } from '../../app/util/object-store/file-store';
 import * as objectStore from '../../app/util/object-store';
-import * as shapefileUpload from '../../app/middleware/shapefile-upload';
 import { stub } from 'sinon';
-import { RequestHandler } from 'express';
-import multer from 'multer';
-import env from '../../app/util/env';
-
-// Patches mock-aws-s3's mock so that the result of "upload" has an "on" method
-// const S3MockPrototype = Object.getPrototypeOf(new mockAws.S3());
-// const originalUpload = S3MockPrototype.upload;
-// S3MockPrototype.upload = function (...args): mockAws.S3.ManagedUpload {
-//   const result = originalUpload.call(this, ...args);
-//   return { on: (): void => {}, ...result };
-// };
-
-// /**
-//  * Adds stubs to S3 object signing that retain the username from the 'A-userid' parameter.
-//  *
-//  * @returns The URL prefix for use in matching responses
-//  */
-// export function hookSignS3Object(): string {
-//   const prefix = 'https://example.com/s3/signed/';
-//   before(function () {
-//     stub(S3ObjectStore.prototype, 'signGetObject')
-//       .callsFake(async (url, params) => `${prefix}${params['A-userid']}`);
-//   });
-//   after(function () {
-//     (S3ObjectStore.prototype.signGetObject as SinonStub).restore();
-//   });
-//   return prefix;
-// }
-
-// /**
-//  * Gets JSON from the given object store URL.  Uses synchronous functions only suitable for testing.
-//  * If using mock-aws-s3, use getObjectText below
-//  * @param url - the Object store URL to get
-//  * @returns the JSON contents of the file at the given URL
-//  */
-// export async function getJson(url: string):
-// Promise<any> { // eslint-disable-line @typescript-eslint/no-explicit-any
-//   const objectStore = new S3ObjectStore();
-//   const filename = await objectStore.downloadFile(url);
-//   try {
-//     return JSON.parse(fs.readFileSync(filename).toString());
-//   } finally {
-//     fs.unlinkSync(filename);
-//   }
-// }
-
-// /**
-//  * Returns the text contents of the object at the provided URL.  If the object is mocked using
-//  * mock-aws-s3 this is likely to produce better results than `getJson` above.
-//  * @param url - the Object store URL to read
-//  */
-// export async function getObjectText(url: string): Promise<string> {
-//   const objectStore = new S3ObjectStore();
-//   const filename = await objectStore.downloadFile(url);
-//   try {
-//     return fs.readFileSync(filename).toString();
-//   } finally {
-//     fs.unlinkSync(filename);
-//   }
-//   // const contents: GetObjectCommandOutput = await new Promise((resolve, reject) => {
-//   //   void objectStoreForProtocol(url).getObject(url, (err, body) => {
-//   //     if (err) reject(err);
-//   //     else resolve(body);
-//   //   });
-//   // });
-//   // return contents.Body.toString();
-// }
-
-
-
-// const shapefileUploadStub = stub(shapefileUpload, 'default');
-// shapefileUploadStub.callsFake(shapefileUploadMock);
-
-// Override using S3 in tests
-// const dir = tmp.dirSync({ unsafeCleanup: true });
-// const fileStore = new FileStore(dir.name);
-// stub(objectStore, 'defaultObjectStore').callsFake(() => fileStore);
-// stub(objectStore, 'objectStoreForProtocol').callsFake(() => fileStore);
-
-// Override multer middleware
-/**
- * Override shapefile upload middleware code which uses S3 and multer
- * @param req - request
- * @param res - response
- * @param next - next middleware function
- * @returns
- */
-function shapefileUploadMock(): RequestHandler {
-  const storage = multer.diskStorage({
-    destination: (_req, _file, callback) => {
-      // Create a unique temporary directory
-      const tempDir = tmp.dirSync().name;
-
-      console.log(`CDD saving to temp dir ${tempDir}`);
-
-      // Save the file to the temporary directory
-      callback(null, tempDir);
-    },
-    filename: (_req, file, callback) => {
-      // Use the original filename of the uploaded file
-      console.log(`CDD in the filename code - using filename ${file.originalname}`);
-      callback(null, file.originalname);
-    },
-  });
-
-  const upload = multer({
-    storage,
-    limits: {
-      fields: env.maxPostFields, // Maximum number of non-file fields to accept
-      fileSize: env.maxPostFileSize, // Maximum size for shapefiles
-      files: 1, // Maximum number of files to accept
-      parts: env.maxPostFileParts, // Maximum number of multipart parts to accept
-    },
-  });
-
-  console.log('CDD - Using mocked out code');
-
-  // Return the actual multer middleware function
-  return upload.single('shapefile');
-  // return upload.fields([{ name: 'shapefile', maxCount: 1 }]);
-}
-
-stub(shapefileUpload, 'default').callsFake(() => shapefileUploadMock());
-// stub(shapefileUpload, 'default').returns((_req, _res, next) => {
-//   // Resolve the middleware by calling the `next` function immediately
-//   next();
-// });
-
-
 
 /**
  * Causes calls to aws.S3 to return a mock S3 object that stores to a temp dir on the
